Update to multiple nested keys
After testing out different methods, I finally found that moving the if statement before the for loop would cause it not to run when a value that isn't a table was being called. This in turn does the same thing as the previous, but allows for the :BeforeSave() function to be used in any case where the value is not a table. From all my testing I found this to work perfectly.